requirements.txt
streamlit>=1.39,<2.0
torch>=2.2,<3.0
torchvision>=0.17,<1.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
opencv-python>=4.9,<5.0
Pillow>=10.4,<13.0
altair>=5.4,<6.0
pytorch-grad-cam>=1.5.0,<2.0
